tenn_ai/requirements.txt
accelerate==0.24.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite>=0.19.0
astroid==2.15.0
async-timeout==4.0.2
attrs==23.1.0
autopep8==1.7.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
chromadb>=0.4.15
click==8.1.6
clickhouse-connect==0.5.22
colorama==0.4.6
datasets>=2.14.5
diffusers>=0.21.0
dill==0.3.6
distlib==0.3.6
docstring-to-markdown==0.11
docx2txt>=0.8
duckduckgo-search==3.8.5
email-validator>=2.0.0
fastapi==0.101.0
fastapi-users>=12.1.2
fastapi-users-db-sqlalchemy>=6.0.1
filelock==3.12.2
flake8==6.0.0
Flask-SocketIO==5.3.4
flask>=2.3.2
frozenlist==1.3.3
google-cloud-aiplatform>=1.34.0
identify==2.5.24
idna==3.4
importlib-metadata==6.8.0
install==1.3.5
ipython==8.14.0
ipywidgets==8.1.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jupyter_core==5.3.1
jupyterlab_widgets==3.0.8
langchain>=0.0.310
langchain-experimental>=0.0.36
lazy-object-proxy==1.9.0
librosa>=0.10.1
litellm>=0.1.610
llama-index>=0.8.55
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mccabe==0.7.0
multidict==6.0.4
nbformat==5.9.2
networkx==3.1
nodeenv==1.8.0
numpy>=1.25.0
openai>=0.28.1
opencv-python==4.8.0.76
open-interpreter>=0.1.5
packaging==23.0
pandas>=2.0.3
pandas-datareader>=0.10.0
parso==0.8.3
pathtools>=0.1.2
Pillow==9.5.0
platformdirs==3.5.3
pluggy==1.2.0
pre-commit==3.3.3
prophet>=1.1.4
pyautogen>=0.1.6
python-docx>=0.8.11
pycodestyle==2.10.0
pydocstyle==6.2.3
pyflakes==3.0.1
pygame==2.5.1
PyJWT>=2.8.0
pylint==2.17.0
pymongo==4.5.0
pypdf>=3.17.0
PySocks==1.7.1
# pystan==3.7.0
pytesseract>=0.3.10
python-decouple>=3.8
python-engineio==4.5.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.1
python-multipart>=0.0.6
python-socketio==5.8.0
pytoolconfig==1.2.5
PyYAML>=5.0
pyxtension>=1.16.1
regex==2023.6.3
requests==2.31.0
retrying==1.3.4
rope==1.7.0
scikit-learn==1.3.0
sentencepiece==0.1.99
setuptools==68.0.0
six==1.16.0
snowballstemmer==2.2.0
soundfile==0.12.1
sqlalchemy>=2.0.22
st_pages==0.4.3
streamlit-tree-select==0.0.5
streamlit==1.25.0
tabulate>=0.9.0
tenacity>=8.2.2
tiktoken==0.4.0
timeout-decorator==0.5.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
torch==2.1.0
tqdm==4.65.0
transformers>=4.34.0
typing_extensions==4.6.0
ujson==5.7.0
urllib3>=1.26.0
virtualenv==20.23.0
watchdog==3.0.0
websockets>=11.0.3
Werkzeug==2.3.6
whatthepatch==1.0.4
wheel>=0.41.2
whisper>=1.1.10
wrapt==1.15.0
yapf==0.32.0
yarl==1.9.2
yfinance>=0.2.28
zipp==3.17.0